feat: added request-refresh message
The request-refresh message is a new message that mobile clients accept. When
called the host will call a refresh on the componenet which made the request.
This message was created so that SSOs and PowerOns in dashboard cards could be
updated without the user having to close or leave the dashboard. This message is
already available in the Android app and will be released in v3.11 of the iOS app.